[Gen] Reverse self.generator keys and values
Keys used to be the generator filename, with the values being a single trigger or a tuple of triggers for that generator.
Now the keys are single triggers, and the values are the generator for each trigger key.

This makes both storing and retrieving the right generator for a trigger much easier